Feat: Fetch random pokemon from API
First: decided to store URLs once fetched to avoid
re-fetching them on every rerender.

Additionally, keep track of fetched pokemon IDs
so that we can randomly fetch a NEW pokemon.
The remaining Ids are stored in a Ref to avoid
recomputation of the remaining ids based on
the fetchedIds state (which is possible, but
costly)